+++ packages/core/test/request-builder/request/get-orderby.spec.ts
@@ -21,7 +21,12 @@ describe('get orderby', () => {
     expect(
       getQueryParametersForOrderBy([asc(TestEntity.COMPLEX_TYPE_PROPERTY.stringProperty), desc(TestEntity.COMPLEX_TYPE_PROPERTY.int16Property)])
         .orderby
-    ).toBe('A_TestComplexType/StringProperty asc,A_TestComplexType/Int16Property desc');
-    // ).toBe('ComplexTypeProperty/StringProperty asc,ComplexTypeProperty/Int16Property desc');
+    ).toBe('ComplexTypeProperty/StringProperty asc,ComplexTypeProperty/Int16Property desc');
+  });
+
+  it('for complex nested types', () => {
+    expect(getQueryParametersForOrderBy([asc(TestEntity.COMPLEX_TYPE_PROPERTY.complexTypeProperty.stringProperty)]).orderby).toBe(
+      'ComplexTypeProperty/ComplexTypeProperty/StringProperty asc'
+    );
   });
 });