+++ packages/generator/test/complex-type/interface.spec.ts
@@ -7,7 +7,7 @@ describe('interface', () => {
     const actual = complexTypeInterface(complexMeal);
     expect(actual).toEqual({
       kind: StructureKind.Interface,
-      name: 'ComplexMeal',
+      name: 'ComplexMealType',
       isExported: true,
       properties: [
         {
@@ -19,7 +19,7 @@ describe('interface', () => {
         },
         { kind: StructureKind.PropertySignature, name: 'amount', type: 'number', hasQuestionToken: false, docs: ['something something very much'] }
       ],
-      docs: ['ComplexMeal']
+      docs: ['ComplexMealType']
     });
   });
 
@@ -27,7 +27,7 @@ describe('interface', () => {
     const actual = complexTypeInterface(complexMealWithDesert);
     expect(actual).toEqual({
       kind: StructureKind.Interface,
-      name: 'ComplexMealWithDesert',
+      name: 'ComplexMealWithDesertType',
       isExported: true,
       properties: [
         {
@@ -39,7 +39,7 @@ describe('interface', () => {
         },
         { kind: StructureKind.PropertySignature, name: 'amount', type: 'number', hasQuestionToken: false, docs: ['something something very much'] }
       ],
-      docs: ['ComplexMealWithDesert']
+      docs: ['ComplexMealWithDesertType']
     });
   });
 });